Issued a certificate. Enabled HTTPs. Created find category by ID use case. Expense update and delete now call Portfolio Manager and verify if Digital User is the owner. Both list use case for Expense and Expense Category now send the IDs (external IDs) to Portfolio Manager.

+++ expense-tracker-core/src/main/java/com/portfolio/expensetracker/usecases/expense/DeleteUseCase.java
@@ -2,26 +2,45 @@ package com.portfolio.expensetracker.usecases.expense;
 
 import com.portfolio.expensetracker.dataprovider.ExpenseDataProvider;
 import com.portfolio.expensetracker.dataprovider.PortfolioManagerDataProvider;
-import com.portfolio.expensetracker.exception.InternalServerErrorException;
+import com.portfolio.expensetracker.dto.portfoliomanager.response.AssetResponse;
+import com.portfolio.expensetracker.exception.AuthorizationFailedException;
 import com.portfolio.expensetracker.security.context.DigitalUser;
 import com.portfolio.expensetracker.util.SecurityUtil;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
 public class DeleteUseCase {
 
+    private final FindByIdUseCase findByIdUseCase;
     private final ExpenseDataProvider dataProvider;
     private final PortfolioManagerDataProvider portfolioManagerDataProvider;
 
     public void execute(Input input) {
-        try {
-            portfolioManagerDataProvider.deleteAsset(input.getJwt(), input.getId());
-        } catch (Exception e) {
-            throw new InternalServerErrorException("Asset could not be deleted in portfolio manager. Please try again.");
+        AssetResponse asset = findByIdUseCase.execute(
+                FindByIdUseCase.Input.builder()
+                        .jwt(input.getJwt())
+                        .id(input.getId())
+                        .build()
+        ).getAsset();
+
+        DigitalUser user = SecurityUtil.getDigitalUser();
+        if (!AssetResponse.PermissionPolicy.OWNER.equals(asset.getPermissionPolicy())) {
+            throw new AuthorizationFailedException(
+                    String.format(
+                            "Digital user %s can not delete expense %s because it is not its owner",
+                            user.getId(),
+                            input.getId()
+                    )
+            );
         }
 
+        portfolioManagerDataProvider.deleteAsset(input.getJwt(), input.getId());
         dataProvider.delete(input.getId());
     }
